requirements.txt
aiogram==2.13
SQLAlchemy==1.4.13
pytest-mock==3.6.0
